+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 2.23.2
 - Fixed error when trying to create cards that evolve/transform in 4 or more turns
 - Fixed Fledgling and Transformer sigil appearing as black boxes when a card evolves/transform in 4 or more turns
+- Fixed Part2VanillaCardCost event not being invoked
 - Evolve and Transformer icons now show support cards that evolve in 4 or more turns
+- Added 'UpdatePlayableCardCosts' event to community patch's Part1CardCostRender and Part2CardCostRender classes
 - Added extension methods for PlayableCard and CardInfo: ProvidesBlueGem, ProvidesGreenGem, ProvidesOrangeGem
 
 # 2.23.1